requirements.txt
mitmproxy>=10.0.0
mysql-connector-python>=8.0.0
aiohttp>=3.8.0
PyYAML>=6.0
